✨(frontend) move ConverseInitializer up in the components tree
As we need to send an askParticipantToJoin through converse, we need to
mount it upper in our components tree.
Also, we have to wait for converse to be fully loaded before requesting
the join.
